Move dataflow_to_sql_converter to common fixture.
