+++ metricflow/test/plan_conversion/test_dataflow_to_sql_plan.py
@@ -63,18 +63,6 @@ def composite_dataflow_to_sql_converter(  # noqa: D
 
 
 @pytest.fixture(scope="session")
-def dataflow_to_sql_converter(  # noqa: D
-    simple_semantic_model: SemanticModel,
-    time_spine_source: TimeSpineSource,
-) -> DataflowToSqlQueryPlanConverter[DataSourceDataSet]:
-    return DataflowToSqlQueryPlanConverter[DataSourceDataSet](
-        column_association_resolver=DefaultColumnAssociationResolver(simple_semantic_model),
-        semantic_model=simple_semantic_model,
-        time_spine_source=time_spine_source,
-    )
-
-
-@pytest.fixture(scope="session")
 def multihop_dataflow_to_sql_converter(  # noqa: D
     multi_hop_join_semantic_model: SemanticModel,
     time_spine_source: TimeSpineSource,
